fix incorrect page count
Happened because at the end of each successful status code 200 if-loop, page gets encremented by 1, so that the next while loop will begin processing the next page.
the if branch that deals with the 301 redirects that literotica gives when you give it an incorrect page index didn't decrement the page counter; hense incorrect page number reporting in the category finished message.