Add a function to parse literotica storyurls from a given HTML string and add a basic (but functional) LE story URL regexp
Fanficfare's results from '-n', which is meant to extract all interesting URLs from a given URL (by interesting, I mean something that fff thinks is a story or an author page, which  contains stories to be downloaded), were too unrelyable.
While that switch pretty much "does what it says on the box", it's not what I'm looking for here; I want story URLs and nothing else when parsing category pages.
It was also returning inconsistent numbers of URLs for each page, which makes me think something is... Off, in that part of fff's code.
So now we have extract_storyurls, who's job it is to... Uh... Yeah. Get storyurls.
I'll probably write a similar function for authorurls, so I can add similar functionality for authors like we have categories (E.G. Create a file with the author names you want larc to download all stories for, one per line), as well as one to parse 'new' listings and follow what categories and authors are specified there and only get stories that match those restrictions.