Fix HUGE bug which made larc infinitly retrieve the last category page
This happened because I forgot to add allow_redirects=False to the requests.get call; that line was originally only a head request to check the status code of the generated URL.
Literotica actually redirects you with a 301 to the last page of results for the given category, and the get call was just following that redirect, infinitely adding the URLs from the last page to the deque.
The if branch that's meant to catch those 301s never triggered, because requests (without being told not to), followed the redirect and returned that page.