Get '--download-new' option working correctly; fix how larc determines the last page of new stories; fix some incorrect variable names
The parse_new_stories function was failing to parse category urls after perfectly doing so with story urls. I'm using the 'class' tag literotica puts on story and category links, and everything worked correctly in the console, so I was baffled.
For some reason, making a copy of the parse tree and running each find_all on their own instance of that tree fixed this issue; this makes me think this is (a) either a beautiful soup "feature" that isn't documented or (b) a bug, I'm thinking the ladder.
It's probably a really bad way to fix this particular issue, but it works, so I'm sticking with the "make a copy" method for now.

I also fixed how larc determines when it's reached the last new stories page; with individual category listings it returns a 301 and a redirect to the last valid page of results when you request a page that is beyond the valid range.
With new stories, it will just return the last page of stories infinitely. So n_previous gets set to the last list of storytuples returned by parse_new_stories if the page being processed is greater than 1, and instead of checking for a 301 http status code, we check if n (the just-downloaded storytuples) is the same as the previously downloaded list.