Fix bug in 'category' mode that would set the finished_retrieving_urls thread event after category workers were started, resulting in no download workers ever actually downloading anything
I basically "fix" this by running a while loop that checks if there are any category workers running, and stops if there aren't, which runs the event-setting code below; the loop just "continues" endlessly.
I have the vague feeling that this is somehow horrible and that there is a better way that I can't find, but my only other ideas (a specific thread to do this) seem just as bad, if not worse because of os overhead.
Also, one major problem this exaserbates is that while this loop is running in the main thread, inside a try / except block the KeyboardInterrupt won't trigger (if raised) until this code has finished running. This was an issue originally, but I'm making it stay an issue at least until the category worker threads have exited. Will have to fix that soon.